+++ CHANGELOG.md
@@ -24,6 +24,7 @@
   Starting from version 1.6.0 [pylint_odoo](https://pypi.python.org/pypi/pylint-odoo/) user `eslint`
 - removed unuseful `odoo-install -y`, it does not run apt, so there is no interactive y/n questions
 - pylint default config: enabled *redefined-outer-name* check
+- experimental `odoo-helper tr rate` command, which computes translation rate for addons. Useful in CI.
 - `odoo-helper status`: print versions of fllowing tools used in current project.
   - NodeJS
   - npm